Add C++ model generator to modelina
This commit makes the basis for the C++ models crafted by tweaking the
template generator and updating the corresponding test cases. On top of
this, I had also ran the test cases adding C++ just to make sure we are
completely clean on renaming the templates to Cpp and it passes all the
cases.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>